backend/pyproject.toml
[project]
name = "luminary-backend"
version = "0.1.0"
description = "Luminary — local-first personal knowledge and learning assistant"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.22.1",
    "arize-phoenix>=13.3.0",
    "beautifulsoup4>=4.14.3",
    "bleach>=6.3.0",
    "chardet>=6.0.0.post1",
    "ebooklib>=0.20",
    "fastapi>=0.132.0",
    "faster-whisper>=1.2.1",
    "fsrs>=6.3.0",
    "gliner>=0.2.25",
    "httpx>=0.28.1",
    "keyring>=25.7.0",
    "kuzu>=0.11.3",
    "lancedb>=0.29.2",
    "langchain-core>=1.2.15",
    "langchain-text-splitters>=1.1.1",
    "langfuse>=3.14.5",
    "langgraph>=1.0.9",
    "litellm>=1.81.14",
    "markdown-it-py>=4.0.0",
    "numpy>=2.4.2",
    "openinference-instrumentation-litellm>=0.1.29",
    "opentelemetry-exporter-otlp-proto-grpc>=1.39.1",
    "opentelemetry-instrumentation-fastapi>=0.60b1",
    "opentelemetry-instrumentation-sqlalchemy>=0.60b1",
    "opentelemetry-sdk>=1.39.1",
    "optimum[onnxruntime]>=2.1.0",
    "pydantic-settings>=2.13.1",
    "pymupdf>=1.27.1",
    "python-docx>=1.2.0",
    "python-json-logger>=4.0.0",
    "python-multipart>=0.0.22",
    "sentence-transformers>=5.2.3",
    "sqlalchemy[asyncio]>=2.0.46",
    "sse-starlette>=3.2.0",
    "tree-sitter>=0.25.2",
    "tree-sitter-go>=0.25.0",
    "tree-sitter-javascript>=0.25.0",
    "tree-sitter-python>=0.25.0",
    "tree-sitter-rust>=0.24.0",
    "tree-sitter-typescript>=0.23.2",
    "uvicorn[standard]>=0.41.0",
    "yt-dlp>=2026.3.13",
]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "PL"]
ignore = [
    "PLC0415",   # import-outside-toplevel: lazy-loading pattern is intentional (embedder)
    "PLR0911",   # too-many-return-statements: acceptable for complex service methods
    "PLR0912",   # too-many-branches: acceptable for ingestion/parsing logic
    "PLR0913",   # too-many-arguments: FastAPI route handlers have many injected params
    "PLR0915",   # too-many-statements: acceptable for complex async generators (stream_answer)
    "PLR2004",   # magic-value-comparison: HTTP status codes (200, 400, etc.) are well-known
    "PLW0603",   # global-statement: lazy singleton pattern for engine/session_factory
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["PL"]
"tools/**" = ["PL"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = "-m 'not slow and not e2e'"
pythonpath = ["tests"]
markers = [
    "slow: marks tests as slow (real ML model downloads; excluded from make test fast path)",
    "e2e: requires live backend + Ollama; excluded from make ci",
    "integration_http: ASGITransport upload tests with mocked ML; included in make ci",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "psutil>=6.1.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "reportlab>=4.4.10",
    "ruff>=0.15.2",
    "tiktoken>=0.12.0",
]
